feat(spx-gui): remove gravity、friction、airDrag item
