+++ spx-gui/src/components/editor/map-editor/MapConfigItemWrapper.vue
